Extract attachment actions into new controller
This renames action as:
  #get_attachment to #show
  #get_attachment_as_html to #show_as_html
